+++ src/rpc/methods/eth.rs
@@ -2046,19 +2046,43 @@ pub enum EthGetTransactionByBlockNumberAndIndex {}
 impl RpcMethod<2> for EthGetTransactionByBlockNumberAndIndex {
     const NAME: &'static str = "Filecoin.EthGetTransactionByBlockNumberAndIndex";
     const NAME_ALIAS: Option<&'static str> = Some("eth_getTransactionByBlockNumberAndIndex");
-    const PARAM_NAMES: [&'static str; 2] = ["p1", "p2"];
+    const PARAM_NAMES: [&'static str; 2] = ["block_param", "tx_index"];
     const API_PATHS: ApiPaths = ApiPaths::V1;
     const PERMISSION: Permission = Permission::Read;
 
-    type Params = (EthUint64, EthUint64);
+    type Params = (BlockNumberOrPredefined, EthUint64);
     type Ok = Option<ApiEthTx>;
 
     async fn handle(
-        _ctx: Ctx<impl Blockstore + Send + Sync + 'static>,
-        (_p1, _p2): Self::Params,
+        ctx: Ctx<impl Blockstore + Send + Sync + 'static>,
+        (block_param, tx_index): Self::Params,
     ) -> Result<Self::Ok, ServerError> {
-        // Lotus doesn't support this method (v1.29.0), so do we.
-        Err(ServerError::unsupported_method())
+        let ts = tipset_by_block_number_or_hash(ctx.chain_store(), block_param.into())?;
+
+        let messages = ctx.chain_store().messages_for_tipset(&ts)?;
+
+        let EthUint64(index) = tx_index;
+        let msg = messages.get(index as usize).with_context(|| {
+            format!(
+                "failed to get transaction at index {}: index {} out of range: tipset contains {} messages",
+                index,
+                index,
+                messages.len()
+            )
+        })?;
+
+        let state = StateTree::new_from_root(ctx.store_owned(), ts.parent_state())?;
+
+        let tx = new_eth_tx(
+            &ctx,
+            &state,
+            ts.epoch(),
+            &ts.key().cid()?,
+            &msg.cid(),
+            index,
+        )?;
+
+        Ok(Some(tx))
     }
 }